Modify TensorAccessor to use size_t instead of int64_t

+++ source/neuropods/internal/tensor_accessor.hh
@@ -18,12 +18,12 @@ private:
 public:
     TensorAccessor(T *data, const int64_t *strides) : data_(data), strides_(strides) {}
 
-    TensorAccessor<T, N - 1> operator[](int64_t i)
+    TensorAccessor<T, N - 1> operator[](size_t i)
     {
         return TensorAccessor<T, N - 1>(this->data_ + this->strides_[0] * i, this->strides_ + 1);
     }
 
-    const TensorAccessor<T, N - 1> operator[](int64_t i) const
+    const TensorAccessor<T, N - 1> operator[](size_t i) const
     {
         return TensorAccessor<T, N - 1>(this->data_ + this->strides_[0] * i, this->strides_ + 1);
     }
@@ -40,9 +40,9 @@ private:
 public:
     TensorAccessor(T *data, const int64_t *strides) : data_(data), strides_(strides) {}
 
-    T &operator[](int64_t i) { return this->data_[i]; }
+    T &operator[](size_t i) { return this->data_[i]; }
 
-    const T &operator[](int64_t i) const { return this->data_[i]; }
+    const T &operator[](size_t i) const { return this->data_[i]; }
 };
 
 } // namespace neuropods
